rename: 폴더명 signIn에서 signup으로 변경
